ref(core): Move browser-only types out of core
Relocate browser-specific types that were exported from `@sentry/core`'s
browser entry but never used by core itself.

- `HandlerDataXhr`, `HandlerDataDom`, `HandlerDataHistory`,
  `SentryWrappedXMLHttpRequest` and `SentryXhrData` move to
  `@sentry/browser-utils`, where the actual instrumentation lives.
- `BrowserClientProfilingOptions` moves to `@sentry/browser` (its only
  consumer, the browser `Client`).
