+++ packages/browser-utils/src/types.ts
@@ -1,9 +1,4 @@
-import type {
-  FetchBreadcrumbHint,
-  HandlerDataFetch,
-  SentryWrappedXMLHttpRequest,
-  XhrBreadcrumbHint,
-} from '@sentry/core';
+import type { FetchBreadcrumbHint, HandlerDataFetch, XhrBreadcrumbHint } from '@sentry/core';
 import { GLOBAL_OBJ } from '@sentry/core';
 
 export const WINDOW = GLOBAL_OBJ as typeof GLOBAL_OBJ &
@@ -29,3 +24,50 @@ export type FetchHint = FetchBreadcrumbHint & {
   input: HandlerDataFetch['args'];
   response: Response;
 };
+
+// This should be: null | Blob | BufferSource | FormData | URLSearchParams | string
+// But since not all of those are available in node, we just use `unknown` here for now
+type XHRSendInput = unknown;
+
+export interface SentryWrappedXMLHttpRequest {
+  __sentry_xhr_v3__?: SentryXhrData;
+  __sentry_own_request__?: boolean;
+  // span id for the xhr request
+  __sentry_xhr_span_id__?: string;
+  setRequestHeader?: (key: string, val: string) => void;
+  getResponseHeader?: (key: string) => string | null;
+}
+
+// WARNING: When the shape of this type is changed bump the version in `SentryWrappedXMLHttpRequest`
+export interface SentryXhrData {
+  method: string;
+  url: string;
+  status_code?: number;
+  body?: XHRSendInput;
+  request_body_size?: number;
+  response_body_size?: number;
+  request_headers: Record<string, string>;
+}
+
+export interface HandlerDataXhr {
+  xhr: SentryWrappedXMLHttpRequest;
+  startTimestamp?: number;
+  endTimestamp?: number;
+  error?: unknown;
+  // This is to be consumed by the HttpClient integration
+  virtualError?: unknown;
+}
+
+export interface HandlerDataDom {
+  // TODO: Replace `object` here with a vendored type for browser Events. We can't depend on the `DOM` or `react` TS types package here.
+  event: object | { target: object };
+  name: string;
+  global?: boolean;
+}
+
+export interface HandlerDataHistory {
+  /** The full URL of the previous page */
+  from: string | undefined;
+  /** The full URL of the new page */
+  to: string;
+}
