+++ packages/browser/test/tracing/request.test.ts
@@ -90,7 +90,7 @@ describe('instrumentOutgoingRequests', () => {
   });
 
   it('creates a QUERY XHR span with the QUERY method attribute', () => {
-    let xhrHandler: ((data: utils.HandlerDataXhr) => void) | undefined;
+    let xhrHandler: ((data: browserUtils.HandlerDataXhr) => void) | undefined;
     let requestSpan: utils.Span | undefined;
 
     vi.spyOn(browserUtils, 'addXhrInstrumentationHandler').mockImplementation(handler => {
@@ -117,7 +117,7 @@ describe('instrumentOutgoingRequests', () => {
         setRequestHeader: vi.fn(),
       },
       startTimestamp: Date.now(),
-    } as utils.HandlerDataXhr);
+    } as browserUtils.HandlerDataXhr);
 
     expect(xhrHandler).toBeDefined();
     expect(requestSpan).toBeDefined();
@@ -134,7 +134,7 @@ describe('instrumentOutgoingRequests', () => {
     it('uses the active propagation context for an ignored child span', () => {
       const activeSpan = new utils.SentryNonRecordingSpan();
       const ignoredSpan = new utils.SentryNonRecordingSpan({ dropReason: 'ignored' });
-      let xhrHandler: ((data: utils.HandlerDataXhr) => void) | undefined;
+      let xhrHandler: ((data: browserUtils.HandlerDataXhr) => void) | undefined;
 
       vi.spyOn(browserUtils, 'addXhrInstrumentationHandler').mockImplementation(handler => {
         xhrHandler = handler;
@@ -163,7 +163,7 @@ describe('instrumentOutgoingRequests', () => {
           setRequestHeader: vi.fn(),
         },
         startTimestamp: Date.now(),
-      } as utils.HandlerDataXhr);
+      } as browserUtils.HandlerDataXhr);
 
       expect(xhrHandler).toBeDefined();
       expect(getTraceDataSpy).toHaveBeenCalledWith({
