+++ packages/core/src/browser-exports.ts
@@ -18,11 +18,4 @@ export { spanStreamingIntegration } from './integrations/browserSpanStreaming';
 export { getLocationHref } from './utils/browser';
 export { supportsDOMError, supportsHistory, supportsNativeFetch, supportsReportingObserver } from './utils/supports';
 export type { XhrBreadcrumbData, XhrBreadcrumbHint } from './types/breadcrumb';
-export type {
-  HandlerDataXhr,
-  HandlerDataDom,
-  HandlerDataHistory,
-  SentryXhrData,
-  SentryWrappedXMLHttpRequest,
-} from './types/instrument';
-export type { BrowserClientReplayOptions, BrowserClientProfilingOptions } from './types/browseroptions';
+export type { BrowserClientReplayOptions } from './types/browseroptions';
