+++ packages/core/src/types/instrument.ts
@@ -1,41 +1,6 @@
-// This should be: null | Blob | BufferSource | FormData | URLSearchParams | string
-// But since not all of those are available in node, we just export `unknown` here for now
-
 import type { WebFetchHeaders } from './webfetchapi';
 
-// Make sure to cast it where needed!
-type XHRSendInput = unknown;
-
 export type ConsoleLevel = 'debug' | 'info' | 'warn' | 'error' | 'log' | 'assert' | 'trace';
-
-export interface SentryWrappedXMLHttpRequest {
-  __sentry_xhr_v3__?: SentryXhrData;
-  __sentry_own_request__?: boolean;
-  // span id for the xhr request
-  __sentry_xhr_span_id__?: string;
-  setRequestHeader?: (key: string, val: string) => void;
-  getResponseHeader?: (key: string) => string | null;
-}
-
-// WARNING: When the shape of this type is changed bump the version in `SentryWrappedXMLHttpRequest`
-export interface SentryXhrData {
-  method: string;
-  url: string;
-  status_code?: number;
-  body?: XHRSendInput;
-  request_body_size?: number;
-  response_body_size?: number;
-  request_headers: Record<string, string>;
-}
-
-export interface HandlerDataXhr {
-  xhr: SentryWrappedXMLHttpRequest;
-  startTimestamp?: number;
-  endTimestamp?: number;
-  error?: unknown;
-  // This is to be consumed by the HttpClient integration
-  virtualError?: unknown;
-}
 
 interface SentryFetchData {
   method: string;
@@ -66,24 +31,10 @@ export interface HandlerDataFetch {
   headers?: WebFetchHeaders;
 }
 
-export interface HandlerDataDom {
-  // TODO: Replace `object` here with a vendored type for browser Events. We can't depend on the `DOM` or `react` TS types package here.
-  event: object | { target: object };
-  name: string;
-  global?: boolean;
-}
-
 export interface HandlerDataConsole {
   level: ConsoleLevel;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   args: any[];
-}
-
-export interface HandlerDataHistory {
-  /** The full URL of the previous page */
-  from: string | undefined;
-  /** The full URL of the new page */
-  to: string;
 }
 
 export interface HandlerDataError {
