+++ packages/replay-internal/test/types.ts
@@ -1,3 +1,3 @@
-import type { HandlerDataDom } from '@sentry/core';
+import type { HandlerDataDom } from '@sentry/browser-utils';
 
 export type DomHandler = (data: HandlerDataDom) => void;
