+++ packages/replay-internal/test/unit/coreHandlers/handleNetworkBreadcrumbs.test.ts
@@ -3,13 +3,8 @@
  */
 
 import '../../utils/mock-internal-setTimeout';
-import type {
-  Breadcrumb,
-  BreadcrumbHint,
-  FetchBreadcrumbHint,
-  SentryWrappedXMLHttpRequest,
-  XhrBreadcrumbHint,
-} from '@sentry/core';
+import type { Breadcrumb, BreadcrumbHint, FetchBreadcrumbHint, XhrBreadcrumbHint } from '@sentry/core';
+import type { SentryWrappedXMLHttpRequest } from '@sentry/browser-utils';
 import { SENTRY_XHR_DATA_KEY } from '@sentry/browser-utils';
 import { beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
 import { NETWORK_BODY_MAX_SIZE } from '../../../src/constants';
